Deb || reverted change

+++ src/transformer/RMSNorm.py
@@ -10,7 +10,7 @@ class RMSNorm(nn.Module):
     # @torch.compile
     def forward(self, hidden_state: torch.Tensor):
         input_dtype = hidden_state.dtype
-        hidden_state = hidden_state.to(torch.float16)
+        hidden_state = hidden_state.to(torch.float32)
         variance = hidden_state.pow(2).mean(-1, keepdim=True)
         hidden_state = hidden_state*torch.rsqrt(variance+self.variance_epsilon)
         return self.weight*hidden_state.to(input_dtype)
